% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{apply_color_to_aa}
\alias{apply_color_to_aa}
\title{Coloring function that applies a gtseq palette}
\usage{
apply_color_to_aa(
  palette = c("Chemistry", "Shapely", "Zappo", "Taylor", "LETTER", "Hydrophobicity")
)
}
\arguments{
\item{palette}{name of the palette}
}
\value{
a coloring function based on chosen palette
}
\description{
Coloring function that applies a gtseq palette
}
\examples{

apply_color_to_aa("Chemistry")(c("A", "K", "L", "-"))

}
