% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{generate_breaks}
\alias{generate_breaks}
\title{takes a vector as input, generates breaks every n times}
\usage{
generate_breaks(x, width = 5)
}
\arguments{
\item{x}{a vector of column names}

\item{width}{the distance between each break}
}
\value{
a vector of column names at every \code{width} position
}
\description{
Give the table's position columns, treat as a numerical axis, and generates breaks using \code{scales::breaks_width}
}
\examples{

library(dplyr)
cnames <- msa |> split_sequences(seq_column = seq) |> dplyr::select(dplyr::contains("pos_")) |> colnames()
generate_breaks(cnames)
}
