% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{Get conensus frequency}
\usage{
get_consensus(x)
}
\arguments{
\item{x}{a vector, ideally of interesting biological sequences}
}
\description{
returns a 1-length named character vector of the most frequent amino acid, where:
\itemize{
\item the value is the frequency
\item the name is amino acid
}
}
\details{
Does not handle ties (one of the ties will be returned)
}
\examples{

get_consensus(c("dog", "dog", "cat"))

}
