% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa}
\alias{msa}
\title{Multiple sequence alignment data}
\format{
\subsection{\code{msa}}{

A tibble
\describe{
\item{group}{Family of coronavirus}
\item{name}{Name of strain}
\item{start}{Start position of sequence}
\item{seq}{sequences}
...
}
}
}
\source{
\url{https://doi.org/10.1038/s41467-021-21968-w}
}
\usage{
msa
}
\description{
Taken from Figure 5 of Wang 2021 “A Conserved Immunogenic and Vulnerable Site on the Coronavirus Spike Protein Delineated by Cross-Reactive Monoclonal Antibodies.” Nature Communications 12 (1): 1715.
}
\keyword{datasets}
