% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_aa}
\alias{palette_aa}
\title{A series of color palettes for amino acid visualization}
\format{
\subsection{\code{palette_aa}}{

A dtibble
\describe{
\item{symbol}{Amino acid symbol}
\item{other}{colors for each palette}
...
}
}
}
\usage{
palette_aa
}
\description{
A series of color palettes for amino acid visualization
}
\keyword{datasets}
