% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{split_sequences}
\alias{split_sequences}
\title{Prepares sequence data frame for gt visualization}
\usage{
split_sequences(df, seq_column, names_prefix = "pos_")
}
\arguments{
\item{df}{expects 1 row per sequence.}

\item{seq_column}{the column that contains sequence data}

\item{names_prefix}{a prefix to add to the resulting columns. Useful for column selection in downstream analysis with gt}
}
\description{
Splits the sequence column into 1 column per character
}
\examples{

msa |> split_sequences(seq_column = seq)
msa |> split_sequences(seq_column = seq, names_prefix = "seq_")

}
