% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{create_html_vertical_bar}
\alias{create_html_vertical_bar}
\title{Create a Simple html Bar For a Bar Chart}
\usage{
create_html_vertical_bar(
  bar_height = 0.5,
  bar_color = "#4a6fa5",
  parent_div_height = "50px",
  parent_div_width = "100\%"
)
}
\arguments{
\item{bar_height}{a number between 0 and 1 that sets the height of the bar as a percentage of the total height of the container (the parent) div}

\item{bar_color}{color of the bars}

\item{parent_div_height}{valid css unit for height of the parent div}

\item{parent_div_width}{valid css unit for width of the parent div}
}
\description{
arent div is set to 100\% width and height
}
\examples{

create_html_vertical_bar(0.5)
}
