% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{get_consensus_return_bar}
\alias{get_consensus_return_bar}
\title{gets the consensus of the vector x, and returns an html bar}
\usage{
get_consensus_return_bar(x)
}
\description{
gets the consensus of the vector x, and returns an html bar
}
