% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtseq.R
\name{gtseq}
\alias{gtseq}
\title{function that takes a dataframe, with sequences in one column, and creates a
gt table}
\usage{
gtseq(
  df,
  seq_column,
  names_prefix = "pos_",
  rowname_col = NULL,
  groupname_col = dplyr::group_vars(df),
  process_md = FALSE,
  caption = NULL,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  auto_align = TRUE,
  id = NULL,
  locale = getOption("gt.locale"),
  row_group.sep = getOption("gt.row_group.sep", " - ")
)
}
\value{
a gt table object
}
\description{
\itemize{
\item splits the seq column into 1 per element
\item applies a color palette to amino acids
}
}
\examples{

msa |> gtseq(seq_column = seq)
}
